работа на дитейлом
перенес создание элементов в отдельный вью фаил
