+++ TODO/Screen/TaskDetail/TaskDetailViewController.swift
@@ -8,360 +8,30 @@
 
 import UIKit
 
-class TaskDetailViewController: UIViewController, UITableViewDelegate {
+class TaskDetailViewController: UIViewController, TaskDetailDelegate {
     
-    //MARK: - VARIABLES
-    var scrollView = UIScrollView()
-    var detailContentView = UIView()
-    var doneButton = UIButton(type: .system)
-    var taskNameTitleLabel = UILabel()
-    var taskNameTextView = UITextView()
-    var taskCreationDateTitleLabel = UILabel()
-    var taskCreationDateLabel = UILabel()
-    var taskDateTitleLabel = UILabel()
-    var taskDateTextField = UITextField()
-    var taskDescriptionTitleLabel = UILabel()
-    var taskDescriptionTextView = UITextView()
-    var checkBlurView: UIVisualEffectView!
-    var toolBarView: UIView!
-    var toolBarStackView: UIStackView!
-    var addCheckButton = UIButton(type: .system)
-    var addCheckElementTextField = UITextField()
-    let checkListTableView = UITableView()
-    
+    var taskDetailView = TaskDetailView()
     var task: Task = Task()
-    let dateFormatter = Main.instance.dateFormatter
-    var notificationPicker = UIDatePicker()
     var router: BaseRouter?
-    let notificationService = Main.instance.notificationService
-    private var currentTheme : String?
-    let minDate = Calendar.current.date(byAdding: .minute, value: 2, to: Date())
-    var bottomAncherConstraint: CGFloat?
     
-    //MARK: - LABEL FACTORY
-    func labelFactory(lab: UILabel, text: String, color: UIColor) -> UILabel {
-        var label = lab
-        label = UILabel()
-        label.translatesAutoresizingMaskIntoConstraints = false
-        label.text = text
-        label.font = UIFont(name: "HelveticaNeue", size: 17)
-        label.textColor = color
-        detailContentView.addSubview(label)
-        return label
+    override func loadView() {
+        taskDetailView.delegate = self
+        view = taskDetailView
+        taskDetailView.task = task
+        taskDetailView.updateData()
     }
     
-    //MARK: - LIFE CYCLE
     override func viewDidLoad() {
         super.viewDidLoad()
-        bottomAncherConstraint = view.frame.height/1.2
-        let theme = Main.instance.themeService.getTheme()
-        view.applyGradient(colours: [theme.backgroundColor, .mainBackground], startX: 0.5, startY: -1.2, endX: 0.5, endY: 0.7)
         router = BaseRouter(viewController: self)
-        notificationPicker.minimumDate = minDate
-        uiSetUp()
-        constrainsInit()
-        view.addTapGestureToHideKeyboard()
     }
     
     override func viewWillAppear(_ animated: Bool) {
-        changeTheme()
+        taskDetailView.changeTheme()
     }
     
-    //MARK: - UI SET UP
-    func uiSetUp(){
-        
-        scrollView.translatesAutoresizingMaskIntoConstraints = false
-        view.addSubview(scrollView)
-        detailContentView.translatesAutoresizingMaskIntoConstraints = false
-        scrollView.addSubview(detailContentView)
-        
-        doneButton.setTitle("Готово", for: .normal)
-        doneButton.translatesAutoresizingMaskIntoConstraints = false
-        doneButton.tintColor = .systemYellow
-        doneButton.titleLabel?.font = UIFont(name: "HelveticaNeue", size: 17)
-        detailContentView.addSubview(doneButton)
-        doneButton.addTarget(self,
-                             action: #selector(handleDoneTouchUpInside),
-                             for: .touchUpInside)
-        
-        taskNameTitleLabel = labelFactory(lab: self.taskNameTitleLabel, text: "Задача", color: .systemGray)
-        
-        taskNameTextView.translatesAutoresizingMaskIntoConstraints = false
-        taskNameTextView.text = (task.name != "" ? task.name : "Напишите название задачи")
-        taskNameTextView.backgroundColor = UIColor.clear
-        taskNameTextView.isEditable = true
-        taskNameTextView.isScrollEnabled = true
-        taskNameTextView.textColor = .systemYellow
-        taskNameTextView.contentInsetAdjustmentBehavior = .automatic
-        taskNameTextView.font = UIFont(name: "HelveticaNeue", size: 17)
-        taskNameTextView.keyboardAppearance = .dark
-        detailContentView.addSubview(taskNameTextView)
-        
-        taskCreationDateTitleLabel = labelFactory(lab: self.taskCreationDateTitleLabel, text: "Дата регестрации задачи:", color: .systemGray)
-        
-        taskCreationDateLabel = labelFactory(lab: self.taskCreationDateLabel, text: dateFormatter.string(from: task.creationDate), color: .systemYellow)
-        
-        taskDateTitleLabel = labelFactory(lab: self.taskDateTitleLabel, text: "Дата уведомления задачи:", color: .systemGray)
-        
-        taskDateTextField.translatesAutoresizingMaskIntoConstraints = false
-        taskDateTextField.attributedPlaceholder = .init(attributedString: NSAttributedString(string: "Дата уведомления не назначена", attributes: [NSAttributedString.Key.font:UIFont.boldSystemFont(ofSize: 17), NSAttributedString.Key.foregroundColor: UIColor.systemYellow]))
-        taskDateTextField.text = task.notificationDate
-        taskDateTextField.textColor = .systemYellow
-        taskDateTextField.font = UIFont(name: "HelveticaNeue", size: 17)
-        taskDateTextField.inputView = notificationPicker
-        taskDateTextField.inputAccessoryView = makeToolBarNotificationsDetail()
-        taskDateTextField.clearsOnBeginEditing = true
-        if #available(iOS 13.4, *) {notificationPicker.preferredDatePickerStyle = .wheels}
-        taskDateTextField.keyboardAppearance = .dark
-        detailContentView.addSubview(taskDateTextField)
-        
-        taskDescriptionTitleLabel = labelFactory(lab: self.taskDescriptionTitleLabel, text: "Описание задачи:", color: .systemGray)
-        
-        taskDescriptionTextView.translatesAutoresizingMaskIntoConstraints = false
-        taskDescriptionTextView.backgroundColor = UIColor.clear
-        taskDescriptionTextView.contentInsetAdjustmentBehavior = .automatic
-        taskDescriptionTextView.isEditable = true
-        taskDescriptionTextView.isScrollEnabled = true
-        taskDescriptionTextView.text = (task.taskDescription != "" ? task.taskDescription : "Напишите описание задачи")
-        taskDescriptionTextView.textAlignment = .left
-        taskDescriptionTextView.textColor = .systemYellow
-        taskDescriptionTextView.font = UIFont(name: "HelveticaNeue", size: 17)
-        taskDescriptionTextView.keyboardAppearance = .dark
-        detailContentView.addSubview(taskDescriptionTextView)
-        
-        checkBlurView = UIVisualEffectView()
-        checkBlurView.translatesAutoresizingMaskIntoConstraints = false
-        checkBlurView.backgroundColor = .clear
-        checkBlurView.effect = UIBlurEffect(style: .systemThinMaterial)
-        detailContentView.addSubview(checkBlurView)
-        
-        toolBarView = UIView()
-        toolBarView.translatesAutoresizingMaskIntoConstraints = false
-        toolBarView.backgroundColor = .systemYellow
-        checkBlurView.contentView.addSubview(toolBarView)
-        
-        toolBarStackView = UIStackView()
-        toolBarStackView.translatesAutoresizingMaskIntoConstraints = false
-        toolBarStackView.axis = .horizontal
-        toolBarStackView.alignment = .fill
-        toolBarStackView.distribution = .fill
-        toolBarView.addSubview(toolBarStackView)
-        
-        addCheckElementTextField.translatesAutoresizingMaskIntoConstraints = false
-        addCheckElementTextField.attributedPlaceholder = .init(attributedString: NSAttributedString(string: "Добавить элемент", attributes: [NSAttributedString.Key.font:UIFont.boldSystemFont(ofSize: 17), NSAttributedString.Key.foregroundColor: UIColor.systemYellow]))
-        addCheckElementTextField.text = ""
-        addCheckElementTextField.textColor = .systemYellow
-        addCheckElementTextField.font = UIFont(name: "HelveticaNeue", size: 15)
-        addCheckElementTextField.clearsOnBeginEditing = true
-        addCheckElementTextField.backgroundColor = .mainBackground
-        addCheckElementTextField.borderStyle = .roundedRect
-        addCheckElementTextField.keyboardAppearance = .dark
-        toolBarStackView.addArrangedSubview(addCheckElementTextField)
-        
-        addCheckButton.translatesAutoresizingMaskIntoConstraints = false
-        addCheckButton.setImage(UIImage(systemName: "plus.app"), for: .normal)
-        addCheckButton.tintColor = .mainBackground
-        addCheckButton.titleLabel?.font = UIFont(name: "HelveticaNeue", size: 22)
-        addCheckButton.addTarget(self,
-                             action: #selector(checkTablePlusAction),
-                             for: .touchUpInside)
-        toolBarStackView.addArrangedSubview(addCheckButton)
+    func taskDetailDismiss(){
+        router?.dismiss()
+    }
 
-        checkListTableView.backgroundColor = .mainBackground
-        checkListTableView.translatesAutoresizingMaskIntoConstraints = false
-        detailContentView.addSubview(checkListTableView)
-        checkListTableView.register(CheckTableViewCell.self, forCellReuseIdentifier: "cell")
-        checkListTableView.rowHeight = 48
-        checkListTableView.dataSource = self
-        checkListTableView.delegate = self
-    }
-    
-    //MARK: - ACTIONS
-    @objc func handleDoneTouchUpInside(){
-        task.name = (taskNameTextView.text == "Напишите название задачи" ? "" : taskNameTextView.text)
-        task.taskDescription = (taskDescriptionTextView.text == "Напишите описание задачи" ? "" : taskDescriptionTextView.text)
-            
-        if task.notificationDate == "" {
-            if taskDateTextField.text != "" {
-                print("New notif")
-                task.notificationDate = taskDateTextField.text
-                task.notificationID = UUID().uuidString
-                task.notificationID = notificationService.sendNotificationRequest(task: task)
-            }
-        } else {
-            if taskDateTextField.text != "" {
-                if taskDateTextField.text != task.notificationDate {
-                    print("Update notif")
-                    task.notificationDate = taskDateTextField.text
-                    task.notificationID = notificationService.updateNotificationRequest(task: task, notificationIdentifier: task.notificationID!)
-                }
-            } else {
-                print("Delete notif")
-                notificationService.deleteNotificationRequest(notificationIdentifier: (task.notificationID)!)
-                task.notificationDate = ""
-                task.notificationID = ""
-                taskDateTextField.text = ""
-            }
-        }
-        
-        task.notificationDate = taskDateTextField.text
-        
-        try? Main.instance.updateTask(task: task)
-        router?.dismiss(animated: true, completion: nil)
-    }
-    
-    @objc func chooseNotificationAction() {
-        UNUserNotificationCenter.current().requestAuthorization(options: [.alert, .sound, .badge]) { granted, error in
-            guard granted else {
-                print("Разрешение не получено")
-                return
-            }
-        }
-        taskDateTextField.text = dateFormatter.string(from: notificationPicker.date)
-        view.endEditing(true)
-    }
-    
-    @objc func checkTablePlusAction(){
-        let id = (task.checkList.max()?.id ?? 0) + 1
-        print("plus id: ", id)
-        let checkMark = CheckMark(id: id, title: addCheckElementTextField.text ?? "", isMarkSelected: false)
-        task.checkList.append(checkMark)
-        addCheckElementTextField.text = ""
-        addCheckElementTextField.becomeFirstResponder()
-        checkListTableView.reloadData()
-    }
-    
-    //MARK: - CONSTRIAINTS
-    func constrainsInit(){
-        NSLayoutConstraint.activate([
-            
-            scrollView.centerXAnchor.constraint(equalTo: view.centerXAnchor),
-            scrollView.widthAnchor.constraint(equalTo: view.widthAnchor),
-            scrollView.topAnchor.constraint(equalTo: view.topAnchor),
-            scrollView.bottomAnchor.constraint(equalTo: view.bottomAnchor),
-            
-            detailContentView.centerXAnchor.constraint(equalTo: scrollView.centerXAnchor),
-            detailContentView.widthAnchor.constraint(equalTo: scrollView.widthAnchor),
-            detailContentView.topAnchor.constraint(equalTo: scrollView.topAnchor),
-            detailContentView.bottomAnchor.constraint(equalTo: scrollView.bottomAnchor),
-            
-            doneButton.topAnchor.constraint(equalTo: detailContentView.topAnchor, constant: 12),
-            doneButton.rightAnchor.constraint(equalTo: detailContentView.rightAnchor, constant: -12),
-            
-            taskNameTitleLabel.topAnchor.constraint(equalTo: detailContentView.topAnchor, constant: 20),
-            taskNameTitleLabel.leftAnchor.constraint(equalTo: detailContentView.leftAnchor, constant: 12),
-            
-            taskNameTextView.topAnchor.constraint(equalTo: taskNameTitleLabel.topAnchor, constant: 20),
-            taskNameTextView.bottomAnchor.constraint(equalTo: taskCreationDateTitleLabel.topAnchor, constant: 2),
-            taskNameTextView.leftAnchor.constraint(equalTo: detailContentView.leftAnchor, constant: 8),
-            taskNameTextView.rightAnchor.constraint(equalTo: detailContentView.rightAnchor, constant: -12),
-            
-            taskCreationDateTitleLabel.topAnchor.constraint(equalTo: taskNameTextView.topAnchor, constant: 48),
-            taskCreationDateTitleLabel.leftAnchor.constraint(equalTo: detailContentView.leftAnchor, constant: 12),
-            
-            taskCreationDateLabel.topAnchor.constraint(equalTo: taskCreationDateTitleLabel.topAnchor, constant: 28),
-            taskCreationDateLabel.leftAnchor.constraint(equalTo: detailContentView.leftAnchor, constant: 12),
-            
-            taskDateTitleLabel.topAnchor.constraint(equalTo: taskCreationDateLabel.topAnchor, constant: 36),
-            taskDateTitleLabel.leftAnchor.constraint(equalTo: detailContentView.leftAnchor, constant: 12),
-            
-            taskDateTextField.topAnchor.constraint(equalTo: taskDateTitleLabel.topAnchor, constant: 28),
-            taskDateTextField.leftAnchor.constraint(equalTo: detailContentView.leftAnchor, constant: 12),
-            
-            taskDescriptionTitleLabel.topAnchor.constraint(equalTo: taskDateTextField.topAnchor, constant: 36),
-            taskDescriptionTitleLabel.leftAnchor.constraint(equalTo: detailContentView.leftAnchor, constant: 12),
-            
-            taskDescriptionTextView.topAnchor.constraint(equalTo: taskDescriptionTitleLabel.topAnchor, constant: 20),
-            taskDescriptionTextView.bottomAnchor.constraint(equalTo: checkBlurView.topAnchor, constant: -4),
-            taskDescriptionTextView.leftAnchor.constraint(equalTo: detailContentView.leftAnchor, constant: 8),
-            taskDescriptionTextView.rightAnchor.constraint(equalTo: detailContentView.rightAnchor, constant: -12),
-            
-            checkBlurView.topAnchor.constraint(equalTo: taskDescriptionTextView.topAnchor, constant: 116),
-            checkBlurView.leftAnchor.constraint(equalTo: detailContentView.leftAnchor, constant: 0),
-            checkBlurView.bottomAnchor.constraint(equalTo: detailContentView.bottomAnchor, constant: -bottomAncherConstraint!),
-            checkBlurView.rightAnchor.constraint(equalTo: detailContentView.rightAnchor, constant: 0),
-            
-            toolBarView.topAnchor.constraint(equalTo: checkBlurView.topAnchor, constant: 0),
-            toolBarView.leftAnchor.constraint(equalTo: checkBlurView.leftAnchor, constant: 0),
-            toolBarView.rightAnchor.constraint(equalTo: checkBlurView.rightAnchor, constant: 0),
-            toolBarView.bottomAnchor.constraint(equalTo: checkBlurView.bottomAnchor, constant: 0),
-            toolBarView.heightAnchor.constraint(equalToConstant: 40),
-            
-            toolBarStackView.topAnchor.constraint(equalTo: toolBarView.topAnchor, constant: 6),
-            toolBarStackView.leftAnchor.constraint(equalTo: toolBarView.leftAnchor, constant: 6),
-            toolBarStackView.rightAnchor.constraint(equalTo: toolBarView.rightAnchor, constant: 6),
-            toolBarStackView.bottomAnchor.constraint(equalTo: toolBarView.bottomAnchor, constant: -6),
-            
-            checkListTableView.topAnchor.constraint(equalTo: toolBarView.bottomAnchor, constant: 0),
-            checkListTableView.leftAnchor.constraint(equalTo: checkBlurView.contentView.leftAnchor, constant: 0),
-            checkListTableView.bottomAnchor.constraint(equalTo: detailContentView.bottomAnchor, constant: 0),
-            checkListTableView.rightAnchor.constraint(equalTo: checkBlurView.contentView.rightAnchor, constant: 0),
-            
-            addCheckButton.widthAnchor.constraint(equalToConstant: 54)
-        ])
-    }
-    
-    //MARK: - CHANGE THEME
-    func changeTheme() {
-        let theme = Main.instance.themeService.getTheme()
-        navigationController?.navigationBar.barTintColor = theme.backgroundColor
-        navigationController?.navigationBar.tintColor = theme.interfaceColor
-        doneButton.tintColor = theme.interfaceColor
-        taskNameTextView.textColor = theme.interfaceColor
-        taskCreationDateLabel.textColor = theme.interfaceColor
-        taskDateTextField.textColor = theme.interfaceColor
-        taskDescriptionTextView.textColor = theme.interfaceColor
-        addCheckElementTextField.textColor = theme.interfaceColor
-        toolBarView.backgroundColor = theme.interfaceColor
-        view.backgroundColor = theme.interfaceColor
-        view.applyGradient(colours: [theme.backgroundColor, .mainBackground], startX: 0.5, startY: -1.2, endX: 0.5, endY: 0.7)
-        addCheckElementTextField.attributedPlaceholder = .init(attributedString: NSAttributedString(string: "Добавить элемент", attributes: [NSAttributedString.Key.font:UIFont.boldSystemFont(ofSize: 17), NSAttributedString.Key.foregroundColor: theme.interfaceColor]))
-        taskDateTextField.attributedPlaceholder = .init(attributedString: NSAttributedString(string: "Дата уведомления не назначена", attributes: [NSAttributedString.Key.font:UIFont.boldSystemFont(ofSize: 17), NSAttributedString.Key.foregroundColor: theme.interfaceColor]))
-    }
-    
-    //MARK: - CHANGE STATE SETTINGS
-    func changeState(state: String) {
-//        self.currentTheme = state
-//        switch Main.instance.state {
-//        case "1":
-//            doneButton.tintColor = .systemYellow
-//            taskNameTextView.textColor = .systemYellow
-//            taskCreationDateLabel.textColor = .systemYellow
-//            taskDateTextField.textColor = .systemYellow
-//            taskDescriptionTextView.textColor = .systemYellow
-//            addCheckElementTextField.textColor = .systemYellow
-//            toolBarView.backgroundColor = .systemYellow
-//            view.backgroundColor = .systemYellow
-//            view.applyGradient(colours: [.vitDarkBrown, .vitBackground], startX: 0.5, startY: -1.2, endX: 0.5, endY: 0.7)
-//            addCheckElementTextField.attributedPlaceholder = .init(attributedString: NSAttributedString(string: "Добавить элемент", attributes: [NSAttributedString.Key.font:UIFont.boldSystemFont(ofSize: 17), NSAttributedString.Key.foregroundColor: UIColor.systemYellow]))
-//            taskDateTextField.attributedPlaceholder = .init(attributedString: NSAttributedString(string: "Дата уведомления не назначена", attributes: [NSAttributedString.Key.font:UIFont.boldSystemFont(ofSize: 17), NSAttributedString.Key.foregroundColor: UIColor.systemYellow]))
-//        case "2":
-//            doneButton.tintColor = .alexeyBackground
-//            taskNameTextView.textColor = .alexeyBackground
-//            taskCreationDateLabel.textColor = .alexeyBackground
-//            taskDateTextField.textColor = .alexeyBackground
-//            taskDescriptionTextView.textColor = .alexeyBackground
-//            addCheckElementTextField.textColor = .alexeyBackground
-//            toolBarView.backgroundColor = .alexeyBackground
-//            view.backgroundColor = .alexeyBackground
-//            view.applyGradient(colours: [.alexeyFog, .vitBackground], startX: 0.5, startY: -1.2, endX: 0.5, endY: 0.7)
-//            addCheckElementTextField.attributedPlaceholder = .init(attributedString: NSAttributedString(string: "Добавить элемент", attributes: [NSAttributedString.Key.font:UIFont.boldSystemFont(ofSize: 17), NSAttributedString.Key.foregroundColor: UIColor.alexeyBackground]))
-//            taskDateTextField.attributedPlaceholder = .init(attributedString: NSAttributedString(string: "Дата уведомления не назначена", attributes: [NSAttributedString.Key.font:UIFont.boldSystemFont(ofSize: 17), NSAttributedString.Key.foregroundColor: UIColor.alexeyBackground]))
-//        case "3":
-//            doneButton.tintColor = .red
-//            taskNameTextView.textColor = .red
-//            taskCreationDateLabel.textColor = .red
-//            taskDateTextField.textColor = .red
-//            taskDescriptionTextView.textColor = .red
-//            addCheckElementTextField.textColor = .red
-//            toolBarView.backgroundColor = .red
-//            view.backgroundColor = .alexLightGray
-//            view.applyGradient(colours: [.alexDarkRed, .vitBackground], startX: 0.5, startY: -1.2, endX: 0.5, endY: 0.7)
-//            addCheckElementTextField.attributedPlaceholder = .init(attributedString: NSAttributedString(string: "Добавить элемент", attributes: [NSAttributedString.Key.font:UIFont.boldSystemFont(ofSize: 17), NSAttributedString.Key.foregroundColor: UIColor.red]))
-//            taskDateTextField.attributedPlaceholder = .init(attributedString: NSAttributedString(string: "Дата уведомления не назначена", attributes: [NSAttributedString.Key.font:UIFont.boldSystemFont(ofSize: 17), NSAttributedString.Key.foregroundColor: UIColor.red]))
-//        default:
-//            break
-//        }
-    }
-    
 }
